+++ examples/shopping_app/lib/route/metadata.g.dart
@@ -194,7 +194,7 @@ RouteDebugInfo shoppingRouteDebugInfo(ShoppingRoutePath route) {
   };
 }
 
-List<Object> shoppingRouteGuards(
+List<RouteGuardBase<ShoppingRoutePath>> shoppingRouteGuards(
   ShoppingRoutePath route,
   Object router,
 ) {